Fix AVX2 and AVX512 builds for mingw
Adds/changes ifdefs in the KeccakP-1600-*.s to support compilation
and linking with mingw-w64, extending earlier work to support macOS.

KeccakP-1600-SnP.h forces usage of the System V AMD64 ABI when built
with mingw to ensure the correct registers are used to pass args
to the assembly code.

ToTargetMakefile.xsl is modified to pass the correct defines to AS
for both macOS and mingw builds.