Add a CI to XKCP, with native as well as cross-compilation tests (compilation and unit tests).
The rationale of the CI script is to:
- Cross-compile all the possible targets
- Run them using Qemu in user mode (for both x86 and native x86) or Wine
  (for native windows mingw cross-compilation).

Even the microcontrollers ARM Cortex-M variants are cross-compiled and
tested (using Qemu in system mode and semi-hosting).

For now, only the 'UnitTests -p' toggle is used as the '-a' toggle takes too
much time.

The CI can be run independently from github runners as it is a dedicated Docker.
Just run (with docker installed of course) at the root folder:

$ CI_VERBOSE=1 make -f Makefile.ci

(since native builds are executed with wine here, users of Mac M1 / ARM
should expect some issues with these: future updates of the CI script could
detect this and avoid Wine in this case. Anyway, CI should work on any x86
plarform).