Fix AVX512 CI: since Qemu is not compatible with AVX512 and since '-march=native' is used, we need to detect at runtime during CI if the running CPU supports AVX512 or not, and adapt the execution script.
This commit also introduces compilation artifacts generation: those are
generated in CI/artifacts.tar.gz. In case of CI errors, because of the
way Docker handles failure, the artifacts are only generated when forcing
the environment variable 'FAILURE_ARTIFACTS=1' (this is helpful to debug
the origin of failure on statically compiled binaries, and see what is the
source of a runtime error).